factor out optimizer configuration for mlm and mbart
